Fix Google Analytics not working on GitHub Pages
I am using a special 'prod' environment to avoid conflicting with an
issue that exists with one of the GitHub Pages gems when used outside
of the GitHub Pages environment. This broke Google Analytics tracking
on GitHub Pages since the environment is set to 'production' on
GitHub Pages. This change causes Google Analytics to be included in
either the 'production' or 'prod' environments.